fix gain precision

+++ trunk-recorder/source.cc
@@ -314,7 +314,7 @@ double Source::get_error() {
 
 /* -- Gain -- */
 
-void Source::set_gain(int r) {
+void Source::set_gain(double r) {
   if (driver == "osmosdr") {
     gain = r;
     cast_to_osmo_sptr(source_block)->set_gain(gain);
@@ -327,7 +327,7 @@ void Source::set_gain(int r) {
   }
 }
 
-void Source::add_gain_stage(std::string stage_name, int value) {
+void Source::add_gain_stage(std::string stage_name, double value) {
   Gain_Stage_t stage = {stage_name, value};
   gain_stages.push_back(stage);
 }
@@ -336,7 +336,7 @@ std::vector<Gain_Stage_t> Source::get_gain_stages() {
   return gain_stages;
 }
 
-void Source::set_gain_by_name(std::string name, int new_gain) {
+void Source::set_gain_by_name(std::string name, double new_gain) {
   if (driver == "osmosdr") {
     cast_to_osmo_sptr(source_block)->set_gain(new_gain, name);
     BOOST_LOG_TRIVIAL(info) << name << " Gain set to: " << cast_to_osmo_sptr(source_block)->get_gain(name);
@@ -359,7 +359,7 @@ int Source::get_gain_by_name(std::string name) {
   return -1;
 }
 
-int Source::get_gain() {
+double Source::get_gain() {
   return gain;
 }
 
@@ -375,7 +375,7 @@ void Source::set_gain_mode(bool m) {
   }
 }
 
-int Source::get_if_gain() {
+double Source::get_if_gain() {
   return if_gain;
 }
